* Move testobjects.go under cdc package.
