+++ pkg/cdc/testobjects.go
@@ -1,4 +1,4 @@
-package test
+package cdc
 
 import (
 	cassdcapi "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
